Upstream version 2.1.1+ds1

# gpg: Signature made sab 13 ott 2012 21:45:14 CEST using DSA key ID 1392B174
# gpg: Good signature from "David Paleino <d.paleino@gmail.com>"
# gpg:                 aka "[jpeg image of size 6708]"
# gpg:                 aka "Hanska <aksnah@gmail.com>"
# gpg:                 aka "David Paleino <david.paleino@poste.it>"
# gpg:                 aka "David Paleino <dapal@debian.org>"

+++ ninja_ide/core/file_manager.py
@@ -1,10 +1,28 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of NINJA-IDE (http://ninja-ide.org).
+#
+# NINJA-IDE is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# any later version.
+#
+# NINJA-IDE is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with NINJA-IDE; If not, see <http://www.gnu.org/licenses/>.
 import os
 import re
 import threading
+import shutil
 
 from PyQt4 import QtCore
 
 from ninja_ide.core import settings
+
 
 #Lock to protect the file's writing operation
 file_store_content_lock = threading.Lock()
@@ -17,11 +35,23 @@ class NinjaIOException(Exception):
     pass
 
 
+class NinjaFileExistsException(Exception):
+    """
+    Try to override existing file without confirmation exception.
+    """
+
+    def __init__(self, filename=''):
+        Exception.__init__(self, 'The file already exists.')
+        self.filename = filename
+
+
 def create_init_file(folderName):
     """Create a __init__.py file in the folder received."""
     if not os.path.isdir(folderName):
         raise NinjaIOException("The destination folder does not exist")
     name = os.path.join(folderName, '__init__.py')
+    if file_exists(name):
+        raise NinjaFileExistsException(name)
     f = open(name, 'w')
     f.flush()
     f.close()
@@ -54,12 +84,13 @@ def create_init_file_complete(folderName):
     fi.close()
 
 
-def create_folder(folderName):
+def create_folder(folderName, add_init_file=True):
     """Create a new Folder inside the one received as a param."""
     if os.path.exists(folderName):
         raise NinjaIOException("The folder already exist")
     os.mkdir(folderName)
-    create_init_file(folderName)
+    if add_init_file:
+        create_init_file(folderName)
 
 
 def create_tree_folders(folderName):
@@ -84,36 +115,47 @@ def file_exists(path, fileName=''):
 def _search_coding_line(txt):
     """Search a pattern like this: # -*- coding: utf-8 -*-."""
     coding_pattern = "coding[:=]\s*([-\w.]+)"
-    return re.search(coding_pattern, txt)
+    pat_coding = re.search(coding_pattern, txt)
+    if pat_coding and unicode(pat_coding.groups()[0]) != 'None':
+        return unicode(pat_coding.groups()[0])
+    return None
+
+
+def get_file_encoding(content):
+    """Try to get the encoding of the file using the PEP 0263 rules
+    search the first or the second line of the file
+    Returns the encoding or the default UTF-8
+    """
+    encoding = None
+    lines_to_check = content.split("\n", 2)
+    for index in range(2):
+        if len(lines_to_check) > index:
+            line_encoding = _search_coding_line(lines_to_check[index])
+            if line_encoding:
+                encoding = line_encoding
+                break
+    #if not encoding is set then use UTF-8 as default
+    if encoding is None:
+        encoding = "UTF-8"
+    return encoding
 
 
 def read_file_content(fileName):
     """Read a file content, this function is used to load Editor content."""
-    if not os.path.exists(fileName):
-        raise NinjaIOException("The file does not exist")
-    if not os.path.isfile(fileName):
-        raise NinjaIOException("%s is not a file" % fileName)
-    f = QtCore.QFile(fileName)
-    if not f.open(QtCore.QIODevice.ReadOnly | QtCore.QIODevice.Text):
-        raise NinjaIOException("%s" % f.errorString())
-    #QIODevice.Text convert \r\n to \n
-    stream = QtCore.QTextStream(f)
-    content = stream.readAll()
-    coding_line = _search_coding_line(content)
-    if coding_line:
-        encoding = coding_line.groups()[0].toUpper()
-        #reset the stream options
-        stream.reset()
-        stream.seek(0)
-        #Set the specific encoding
-        stream.setCodec(encoding)
-        content = stream.readAll()
-    f.close()
-    return unicode(content)
+    try:
+        with open(fileName, mode='rU') as f:
+            content = f.read()
+            encoding = get_file_encoding(content)
+            content.decode(encoding)
+    except IOError, reason:
+        raise NinjaIOException(unicode(reason))
+    return content
 
 
 def get_basename(fileName):
     """Get the name of a file or folder specified in a path."""
+    if fileName.endswith(os.path.sep):
+        fileName = fileName[:-1]
     return os.path.basename(fileName)
 
 
@@ -122,19 +164,22 @@ def get_folder(fileName):
     return os.path.dirname(fileName)
 
 
-def _real_store_file_content(fileName, content):
-    """Function that actually save the content of a file (thread)."""
-    global file_store_content_lock
-    file_store_content_lock.acquire()
+def store_file_content(fileName, content, addExtension=True, newFile=False):
+    """Save content on disk with the given file name."""
+    if fileName == '':
+        raise Exception()
+    ext = (os.path.splitext(fileName)[-1])[1:]
+    if ext == '' and addExtension:
+        fileName += '.py'
+    if newFile and file_exists(fileName):
+        raise NinjaFileExistsException(fileName)
     try:
         f = QtCore.QFile(fileName)
         if not f.open(QtCore.QIODevice.WriteOnly | QtCore.QIODevice.Truncate):
             raise NinjaIOException(f.errorString())
-        #QTextStream detect locales ;)
         stream = QtCore.QTextStream(f)
-        coding_line = _search_coding_line(content)
-        if coding_line:
-            encoding = coding_line.groups()[0].toUpper()
+        encoding = get_file_encoding(content)
+        if encoding:
             stream.setCodec(encoding)
         encoded_stream = stream.codec().fromUnicode(content)
         f.write(encoded_stream)
@@ -142,31 +187,15 @@ def _real_store_file_content(fileName, content):
         f.close()
     except:
         raise
-    finally:
-        file_store_content_lock.release()
-
-
-def store_file_content(fileName, content):
-    """Save content on disk with the given file name."""
-    if fileName == '':
-        raise Exception()
-    ext = (os.path.splitext(fileName)[-1])[1:]
-    if ext == '':
-        fileName += '.py'
-    t = threading.Thread(target=_real_store_file_content,
-                            args=(fileName, content))
-    t.start()
-    #wait until the saver finish
-    t.join()
-    return fileName
+    return os.path.abspath(fileName)
 
 
 def open_project(path):
     """Return a dict structure containing the info inside a folder."""
     if not os.path.exists(path):
-        raise NinjaIOException("The folder already exist")
+        raise NinjaIOException("The folder does not exist")
     d = {}
-    for root, dirs, files in os.walk(path):
+    for root, dirs, files in os.walk(path, followlinks=True):
         d[root] = [[f for f in files
                 if (os.path.splitext(f.lower())[-1]) in \
                 settings.SUPPORTED_EXTENSIONS],
@@ -181,9 +210,10 @@ def open_project_with_extensions(path, extensions):
     if not os.path.exists(path):
         raise NinjaIOException("The folder does not exist")
     d = {}
-    for root, dirs, files in os.walk(path):
+    for root, dirs, files in os.walk(path, followlinks=True):
         d[root] = [[f for f in files
-                if (os.path.splitext(f.lower())[-1]) in extensions],
+                if (os.path.splitext(f.lower())[-1]) in extensions or \
+                '.*' in extensions],
                 dirs]
     return d
 
@@ -199,20 +229,35 @@ def delete_file(path, fileName=None):
         os.remove(path)
 
 
+def delete_folder(path, fileName=None):
+    """Delete the proper folder."""
+    if fileName:
+        path = os.path.join(path, fileName)
+    if os.path.isdir(path):
+        shutil.rmtree(path)
+
+
 def rename_file(old, new):
     """Rename a file, changing its name from 'old' to 'new'."""
     if os.path.isfile(old):
         ext = (os.path.splitext(new)[-1])[1:]
         if ext == '':
             new += '.py'
+        if file_exists(new):
+            raise NinjaFileExistsException(new)
         os.rename(old, new)
         return new
     return ''
 
 
 def get_file_extension(fileName):
-    """Get the file extension in the form of: '.py'"""
-    return (os.path.splitext(fileName.lower())[-1])
+    """Get the file extension in the form of: 'py'"""
+    return os.path.splitext(fileName.lower())[-1][1:]
+
+
+def get_file_name(fileName):
+    """Get the file name, without the extension."""
+    return os.path.splitext(fileName)[0]
 
 
 def get_module_name(fileName):
@@ -237,6 +282,8 @@ def create_path(*args):
 
 def belongs_to_folder(path, fileName):
     """Determine if fileName is located under path structure."""
+    if not path.endswith(os.path.sep):
+        path += os.path.sep
     return fileName.startswith(path)
 
 
@@ -261,6 +308,17 @@ def check_for_external_modification(fileName, old_mtime):
 
 def get_files_from_folder(folder, ext):
     """Get the files in folder with the specified extension."""
-    filesExt = os.listdir(folder)
+    try:
+        filesExt = os.listdir(folder)
+    except:
+        filesExt = []
     filesExt = [f for f in filesExt if f.endswith(ext)]
     return filesExt
+
+
+def is_supported_extension(filename, extensions=None):
+    if extensions is None:
+        extensions = settings.SUPPORTED_EXTENSIONS
+    if os.path.splitext(filename.lower())[-1] in extensions:
+        return True
+    return False
